tests

+++ src/main/java/cz/kamenitxan/jakon/core/model/Dao/AbstractHibernateDao.java
@@ -5,16 +5,16 @@ import org.hibernate.Session;
 import java.io.Serializable;
 import java.util.List;
 
-public class AbstractHibernateDao< T extends Serializable> {
+public class AbstractHibernateDao<T extends Serializable> {
 
-	private Class< T > clazz;
+	private Class<T> clazz;
 
 	public AbstractHibernateDao(Class<T> clazz) {
 		this.clazz = clazz;
 	}
 
-	public T findOne( long id ){
-		return (T) getCurrentSession().get( clazz, id );
+	public T findOne(int id) {
+		return (T) getCurrentSession().get(clazz, id);
 	}
 
 	public List<T> findAll() {
@@ -25,20 +25,21 @@ public class AbstractHibernateDao< T extends Serializable> {
 		return all;
 	}
 
-	public void create( T entity ){
-		getCurrentSession().persist( entity );
+	public void create(T entity) {
+		getCurrentSession().persist(entity);
 	}
 
-	public void update( T entity ){
-		getCurrentSession().merge( entity );
+	public void update(T entity) {
+		getCurrentSession().merge(entity);
 	}
 
-	public void delete( T entity ){
-		getCurrentSession().delete( entity );
+	public void delete(T entity) {
+		getCurrentSession().delete(entity);
 	}
-	public void deleteById( long entityId ) {
-		T entity = findOne( entityId );
-		delete( entity );
+
+	public void deleteById(int entityId) {
+		T entity = findOne(entityId);
+		delete(entity);
 	}
 
 	private Session getCurrentSession() {
